Avoid rebuilding product docs in products by category query
We were essentially reconstructing every product object in the `groupProductsByCategory()` query in `ProductController.java`, which probably wasn't very efficient and _definitely_ was a maintenance issue since we'd have to remember to update this code every time we modified the `Product` schema.

Some searching turned up the existence of the `$$ROOT` system variable which, in groupings, gives you the entire object. Using that appears to make everything work just fine.