Work towards making tests pass.
Making progress towards making all of the tests pass on the client after updating the product list. This is mostly appending to the mock services for the karma tests, but there is some work towards the e2e tests. Currently there are still things that are probably timing issues with the e2e tests, and the karma test coverage isn't great right now, (branches and function coverage is only about 77-80%).