Reduce fragility of shopping list delete Cypress test
The previous tests of deleting an item from the shopping list were fragile (they depended on there being 3 items in the "Other Store" list), and *very* incomplete (they never delete anything but the first item from the first store). This allowed a usability issue to creep in, which I'd like to fix. First, though, I want to add some new tests that fail in the current universe to highlight the issues.

This starts that process by trying to make the existing tests a little more general.