Add `getStoreTabByName()` to `shoppinglist.po.ts`
The Cypress tests currently access all the store tabs by *position*, which is really fragile and will break when we fix the store order to be something more sensible. 

This adds a new `getStoreTabByName()`, which gets the tab by the name of the store. This is still fragile (it will break when we fix the missing apostrophe in "Willie's", for example, but is probably better.

Unfortunately it seems really hard to propagate attributes through the Angular Material `mat-tab` element, so we can't use things like `data-test` or `data-cy`.