Rewrite the failing e2e test
The cypress test for checking behavior when adding a product to the shopping list when the product already existed proved to be fragile, and to have broken with other current changes. The test is now rewritten to guarantee that the product already is in the shopping list by making a new product, and then adding it to the shopping list, which then allows for the correct behavior to be tested.