Remove space in CSS to fix z-index for toolbar
The problem with the scrolling seemed to have been that the toolbar was at the same z-index as the other elements.

There was an earlier attempt to ensure that the toolbar stays above other elements by setting it's z-index to 1. This wasn't working, though, because there was a missing space in `app.component.scss`. Removing this space seems to fix the problem.