Extend Cypress `app.spec.ts` tests to include shopping list
We testing that the pantry and products links in the toolbar went to the right URLs, but there was no test for the shopping list. This adds such a test, and re-arranges things so they seem to actually run now. (For reasons not entirely clear to me, the nesting `it` clause in that spec file didn't appear to be running anywhere.)