Fixed main page inconsistent theming.
Change based on the conclusion of the approved answer of [this stackOverflow post](https://stackoverflow.com/questions/46499768/background-color-does-not-switch-between-light-and-dark-themes-in-angular-materi). It turns out that removing the mat-sidenav (in favor of the mat-navbar compoent) component unattached the main body of the page from the angular material styling, and that the existing themes only work within an angular material component. So this commit wraps the HTML <body></body> within an angular mat-app-background class.