Address warnings about `console.log`
DeepSource doesn't want us to be using `console.log()` (which seems reasonable to avoid). I've removed two calls that were likely old debugging calls, and tried using the `skipcq` comment command to see if DeepSource will then ignore the one call that is being used to log an error. That error should be changed to something like a snackbar alert so the user knows something, but I'll leave this alone for now.