Provide return values for functions with missing conditional cases
DeepSource identified four functions where there were conditionals with missing return cases. This fixes those.

The two in `product-form.po.ts` were pretty straightforward as I could take advantage of the typing that ensures that the `mode` was either `'Add'` or `'EDIT'`.

I think that the situation in `product.service.ts` might have been a straight-up bug, and I just moved the `return` outside the conditional.

I'm not thrilled about my solution in `product.service.mock.ts`, as I just return `null`. Given that this is in the mock service, though, I hope that this will be ok.