Make methods that don't use `this` static
DeepSource recommends marking class methods that don't make use of `this` as being `static`.

This starts the process of addressing that. Basically all of our `.po` files are really just collections of `static` methods, so I'm marking them as such, and replacing calls to them with static calls on the class name.